requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.6
djangorestframework==3.15.1
faiss-cpu==1.8.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langsmith==0.1.63
multidict==6.0.5
numpy==1.26.4
openai==1.30.3
orjson==3.10.3
packaging==23.2
pydantic==2.7.1
pydantic_core==2.18.2
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
scipy==1.13.1
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.1
yarl==1.9.4
